Move Blosc2 chunker code from HDF5 publisher root to HDF5 support
With minimal changes for it to work.
